+++ plivoclient.go
@@ -48,6 +48,7 @@ type Client struct {
 	Profile                 *ProfileService
 	Campaign                *CampaignService
 	MaskingSession          *MaskingSessionService
+	VerifySession           *VerifyService
 }
 
 /*
@@ -118,6 +119,7 @@ func NewClient(authId, authToken string, options *ClientOptions) (client *Client
 	client.Campaign = &CampaignService{client: client}
 	client.Profile = &ProfileService{client: client}
 	client.MaskingSession = &MaskingSessionService{client: client}
+	client.VerifySession = &VerifyService{client: client}
 	return
 }